Minor restructuring of content validation
