+++ src/validation/ContentDataValidator.ts
@@ -1,0 +1,260 @@
+import paths from "path";
+
+import { defined } from "../base/defined";
+
+import { Uris } from "../io/Uris";
+import { ResourceTypes } from "../io/ResourceTypes";
+
+import { Validators } from "./Validators";
+import { ValidationContext } from "./ValidationContext";
+
+import { B3dmValidator } from "../tileFormats/B3dmValidator";
+import { I3dmValidator } from "../tileFormats/I3dmValidator";
+import { PntsValidator } from "../tileFormats/PntsValidator";
+import { CmptValidator } from "../tileFormats/CmptValidator";
+import { GltfValidator } from "../tileFormats/GltfValidator";
+
+import { Content } from "../structure/Content";
+
+import { IoValidationIssues } from "../issues/IoValidationIssue";
+import { ContentValidationIssues } from "../issues/ContentValidationIssues";
+
+/**
+ * A class for validation of the data that is pointed to by a `content.uri`.
+ * 
+ *
+ * @private
+ */
+export class ContentDataValidator {
+  /**
+   * Validate the actual data that is referred to by the URI in the
+   * given content.
+   *
+   * This assumes that the given content has already been validated
+   * to be structurally valid, using the `ContentValidator`.
+   *
+   * @param contentPath The path for the `ValidationIssue` instances
+   * @param content The `Content` object
+   * @param context The `ValidationContext`
+   * @returns A promise that resolves when the validation is finished
+   */
+   static async validateContentData(
+    contentPath: string,
+    content: Content,
+    context: ValidationContext
+  ): Promise<void> {
+    // Validate the uri
+    const uri = content.uri;
+    // TODO: Assuming that absolute URIs should not be checked
+    if (Uris.isAbsoluteUri(uri)) {
+      return;
+    }
+
+    const resourceResolver = context.getResourceResolver();
+    const contentData = await resourceResolver.resolve(uri);
+
+    // When the content data is not defined, this means that
+    // the URI could not be resolved
+    if (!defined(contentData)) {
+      const path = contentPath;
+      const message =
+        `Tile content ${contentPath} refers to URI ${uri}, ` +
+        `which could not be resolved`;
+      const issue = IoValidationIssues.IO_WARNING(path, message);
+      context.addIssue(issue);
+      return;
+    } else {
+      await ContentDataValidator.validateContentDataInternal(
+        contentPath,
+        uri,
+        contentData!,
+        context
+      );
+    }
+  }
+
+  /**
+   * Perform the validation of the given content data.
+   *
+   * The data may be binary data from tile formats (B3DM, GLB...),
+   * or JSON data (for external tilesets or glTF). The method will
+   * try to figure out the actual data type using a lot of guesses
+   * and magic (i.e. magic headers), and try to validate the data.
+   *
+   * If the data causes validation issues, they will be summarized
+   * into a `CONTENT_VALIDATION_ERROR` or `CONTENT_VALIDATION_WARNING`
+   * that is added to the given context.
+   *
+   * If the data type cannot be determined, an `CONTENT_VALIDATION_WARNING`
+   * will be added to the given context.
+   *
+   * @param contentPath The path for the `ValidationIssue` instances.
+   * @param contentUri The URI of the content
+   * @param contentData The buffer containing the actual content data
+   * @param context The `ValidationContext`
+   * @returns A promise that resolves when the validation is finished
+   */
+  private static async validateContentDataInternal(
+    contentPath: string,
+    contentUri: string,
+    contentData: Buffer,
+    context: ValidationContext
+  ): Promise<void> {
+    // Figure out the type of the content data and pass it
+    // to the responsible validator.
+
+    const isGlb = ResourceTypes.isGlb(contentData);
+    if (isGlb) {
+      console.log("Validating GLB: " + contentUri);
+      const dataValidator = new GltfValidator(contentUri);
+      await dataValidator.validateObject(contentData, context);
+      return;
+    }
+
+    const isB3dm = ResourceTypes.isB3dm(contentData);
+    if (isB3dm) {
+      console.log("Validating B3DM: " + contentUri);
+      const dataValidator = new B3dmValidator(contentUri);
+      await dataValidator.validateObject(contentData, context);
+      return;
+    }
+
+    const isI3dm = ResourceTypes.isI3dm(contentData);
+    if (isI3dm) {
+      console.log("Validating I3DM: " + contentUri);
+      const dataValidator = new I3dmValidator(contentUri);
+      await dataValidator.validateObject(contentData, context);
+      return;
+    }
+
+    const isPnts = ResourceTypes.isPnts(contentData);
+    if (isPnts) {
+      console.log("Validating PNTS: " + contentUri);
+      const dataValidator = new PntsValidator(contentUri);
+      await dataValidator.validateObject(contentData, context);
+      return;
+    }
+
+    const isCmpt = ResourceTypes.isCmpt(contentData);
+    if (isCmpt) {
+      console.log("Validating CMPT: " + contentUri);
+      const dataValidator = new CmptValidator(contentUri);
+      await dataValidator.validateObject(contentData, context);
+      return;
+    }
+
+    // When there is no known magic value, then it may be JSON.
+    const isJson = ResourceTypes.isProbablyJson(contentData);
+    if (isJson) {
+      await ContentDataValidator.validateJsonContentData(
+        contentPath,
+        contentUri,
+        contentData,
+        context
+      );
+      return;
+    }
+
+    const path = contentPath;
+    const message =
+      `Tile content ${contentPath} refers to URI ${contentUri}, ` +
+      `for which no tile content type could be determined`;
+    const issue = ContentValidationIssues.CONTENT_VALIDATION_WARNING(
+      path,
+      message
+    );
+    context.addIssue(issue);
+  }
+
+  /**
+   * Perform the validation of the given content data, which already
+   * has been determined to (probably) be JSON data.
+   *
+   * The method will try to figure out the actual data type using
+   * a few guesses, and try to validate the data.
+   *
+   * If the data causes validation issues, they will be summarized
+   * into a `CONTENT_VALIDATION_ERROR` or `CONTENT_VALIDATION_WARNING`
+   * that is added to the given context.
+   *
+   * If the data type cannot be determined, an `CONTENT_VALIDATION_WARNING`
+   * will be added to the given context.
+   *
+   * @param contentPath The path for the `ValidationIssue` instances.
+   * @param contentUri The URI of the content
+   * @param contentData The buffer containing the actual content data
+   * @param context The `ValidationContext`
+   * @returns A promise that resolves when the validation is finished
+   */
+  private static async validateJsonContentData(
+    contentPath: string,
+    contentUri: string,
+    contentData: Buffer,
+    context: ValidationContext
+  ) {
+    // If the data is probably JSON, try to parse it in any case,
+    // and bail out if it cannot be parsed
+    let parsedObject = undefined;
+    try {
+      parsedObject = JSON.parse(contentData.toString());
+    } catch (error) {
+      const issue = IoValidationIssues.JSON_PARSE_ERROR(contentUri, "" + error);
+      context.addIssue(issue);
+      return;
+    }
+
+    // Try to rule out JSON files which will not be validated anyhow
+    const ext = paths.extname(contentUri).toLowerCase();
+    if (ext === "geojson") {
+      const message = `Skipping validation of apparent GeoJson file: ${contentUri}`;
+      const issue = ContentValidationIssues.CONTENT_VALIDATION_WARNING(
+        contentUri,
+        message
+      );
+      context.addIssue(issue);
+      return;
+    }
+
+    // An 'asset' may indicate an external tileset or a glTF...
+    if (defined(parsedObject.asset)) {
+      // When there is a `geometricError` or a `root`,
+      // let's assume that it is an external tileset:
+      if (defined(parsedObject.geometricError) || defined(parsedObject.root)) {
+        console.log("Validating as external tileset: " + contentUri);
+        // Create a new context to collect the issues that are
+        // found in the data. If there are issues, then they
+        // will be stored as the 'internal issues' of a
+        // single content validation issue.
+        const dirName = paths.dirname(contentUri);
+        const derivedContext = context.derive(dirName);
+        const externalValidator = Validators.createDefaultTilesetValidator();
+        await externalValidator.validateObject(parsedObject, derivedContext);
+        const derivedResult = derivedContext.getResult();
+        const issue = ContentValidationIssues.createFrom(
+          contentUri,
+          derivedResult
+        );
+        if (issue) {
+          context.addIssue(issue);
+        }
+        return;
+      }
+
+      // The parsed object has an 'asset', but is no tileset.
+      // Assume that it is a glTF:
+      console.log("Validating glTF: " + contentUri);
+      const gltfValidator = new GltfValidator(contentUri);
+      await gltfValidator.validateObject(contentData, context);
+      return;
+    }
+    const path = contentPath;
+    const message =
+      `Tile content ${contentPath} refers to URI ${contentUri}, which ` +
+      `contains JSON data, but for which no type could be determined`;
+    const issue = ContentValidationIssues.CONTENT_VALIDATION_WARNING(
+      path,
+      message
+    );
+    context.addIssue(issue);
+  }
+}
